fix: removing some unnecessary code
